fix(layout):
i fixed the title, I put the top bar last in the template, it worked ... drawing order counts, but in reverse order .... to be confirmed

i fixed scrolling + zcroll distance +/-= 3*mouse.wheel

i fixed incomplete gutter ( tit extends to the bottom with virtual line numbers

i fiddled with the apparent height width and position of boxes in such a way that the layout should look the way i want. u consider this:

THE BOX MODEL IS ESSENTIALLY FLAWED:
a box of height 0 wil have the start y and end y the same wich should result in a line like fiill, same goes for width, if the height is 1 then the y1 == y0 + 1 which will result indrawing a box of height 2 ....

a box in the vt space is not delimited by zero width lines, but denotes the spots that will be filled. its like the distinction between a fence panel and fence posts. you cannot draw fence posts in the VT terminals, its alll panels and the posts between them, the lines of the grid, we dont have any access to them.